Feature/cli configuration (#151)
* Added interactive configuration option to (cx configure). This mode will
desplay previously entered values for common parameters and prompt you
to enter new ones. Pressing 'enter' skips the operation.

* Fixed liniting issues.

* Fixed linting issues.

* Added (--proxy) option to CLI, previously you could only set the proxy
through the environment viarable HtTP_PROXY. renamed HTTP_PROXY to
CX_HTTP_PROXY. This was done both to bring the environment variable into
line with the rest of the CX environment variables but also to map the
value out to VIPER bindings. The previous name was a system name that
GoLang defined.

* Correcting issue with unused variable.